ui.load_device: Fix threading, needs to be emitted
If we directly call the load_device func, then it just runs in current thread (in this case GUI thread).

This means we need to have the thread itself call the func. How? Well the only way i've seen is via signals. I placed an "input" signal called `device` and connected it to Worker.load_device, this means when I emit to the `device` signal, which lives in the thread, it sends that signal value to load_devices (which is done by the thread, whip-eee!)