Rework cfg to be in config.py, heavily improve
The `Config` class (User Config loader and writer) now loads on-demand rather than immediately on object creation. This allows a config file to be instantiated within the config file, and then be importable for quick and instant use across all project files.

This also now adds the ability to reload the config data by simply clearing one variable. All further calls will then be from a new config load.

The project and system strings and such that were in `cfg` are now in `config.py` with some changes to either variable names, or their contents. Generally they are the same except more refined and with a better way to access them. They are also organised by Directories, System, Project.